requirements.txt
asttokens==2.2.1
backcall==0.2.0
bertopic==0.15.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
comm==0.1.4
Cython==0.29.36
debugpy==1.6.7.post1
decorator==5.1.1
emoji==2.7.0
executing==1.2.0
filelock==3.12.2
fsspec==2023.6.0
gensim==4.3.1
hdbscan==0.8.33
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.1
ipython==8.12.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter-client==8.3.0
jupyter-core==5.3.1
llvmlite==0.40.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
numba==0.57.1
numpy==1.24.4
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
pip==21.1.1
platformdirs==3.10.0
plotly==5.16.0
prompt-toolkit==3.0.39
protobuf==4.24.0
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.16.1
pynndescent==0.5.10
PySastrawi==1.2.0
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==56.0.0
six==1.16.0
smart-open==6.3.0
stack-data==0.6.2
stanza==1.5.0
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.31.0
typing-extensions==4.7.1
tzdata==2023.3
umap-learn==0.5.3
urllib3==2.0.4
wcwidth==0.2.6
zipp==3.16.2
